refactor: restore merged-main architecture gates
