+++ crates/ctx-cli/src/semantic/source_backed_refresh_coordinator/refresh_mode.rs
@@ -1,0 +1,16 @@
+#[derive(Debug, Clone, Copy, Eq, PartialEq)]
+pub(crate) enum SourceBackedRefreshMode {
+    Off,
+    Background,
+    Wait,
+}
+
+impl SourceBackedRefreshMode {
+    pub(super) fn as_str(self) -> &'static str {
+        match self {
+            Self::Off => "off",
+            Self::Background => "background",
+            Self::Wait => "wait",
+        }
+    }
+}
